refact: resolução do docs

+++ app/schema/viagem.py
@@ -1,28 +1,27 @@
-from sqlalchemy import BigInteger, Integer
 from pydantic import BaseModel
 from typing import Optional
 
 class ViagemSchema(BaseModel):
-    id_viagem: BigInteger
-    id_avaliacao: BigInteger
-    id_corrida: BigInteger
-    id_modelo: Integer
-    id_motorista: BigInteger
-    id_pagamento: BigInteger
-    id_passageiro: BigInteger
-    id_servico: Integer
-    id_veiculo: BigInteger
+    id_viagem: int
+    id_avaliacao: int
+    id_corrida: int
+    id_modelo: int
+    id_motorista: int
+    id_pagamento: int
+    id_passageiro: int
+    id_servico: int
+    id_veiculo: int
 
     class Config:
         from_attributes = True
 
 class ViagemUpdateSchema(BaseModel):
-    id_viagem: Optional[BigInteger]
-    id_avaliacao: Optional[BigInteger]
-    id_corrida: Optional[BigInteger]
-    id_modelo: Optional[Integer]
-    id_motorista: Optional[BigInteger]
-    id_pagamento: Optional[BigInteger]
-    id_passageiro: Optional[BigInteger]
-    id_servico: Optional[Integer]
-    id_veiculo: Optional[BigInteger]
+    id_viagem: Optional[int]
+    id_avaliacao: Optional[int]
+    id_corrida: Optional[int]
+    id_modelo: Optional[int]
+    id_motorista: Optional[int]
+    id_pagamento: Optional[int]
+    id_passageiro: Optional[int]
+    id_servico: Optional[int]
+    id_veiculo: Optional[int]
